+++ src/main.cpp
@@ -14,47 +14,6 @@
 #include <Geode/loader/Event.hpp>
 using namespace geode::prelude;
 
-static void onsettingsUpdate() {
-    auto selection_release = GetSettingJsonRead("selection-release");
-    ReleaseSound->Setsound(selection_release.Custom_Sound_Path);
-
-    auto selection_clicks = GetSettingJsonRead("selection-clicks");
-    ClickSound->Setsound(selection_clicks.Custom_Sound_Path);
-    if (ClickJson->hassomedata) {
-
-        Custom_OnClick = selection_clicks.M_Tab == 2;
-        if (selection_clicks.M_Tab == 0) {
-            auto list = ClickJson->GetMemeClicks();
-            auto sound = list.find(selection_clicks.Current_Sound_Meme);
-            if (sound != list.end()) {
-                ClickSoundIndex->SetSounds(sound->second.clicks, "click-volume","selection-clicks");
-            }
-        } else {
-            auto list = ClickJson->GetUsefulClicks();
-            auto sound = list.find(selection_clicks.Current_Sound_Useful);
-            if (sound != list.end()) {
-                ClickSoundIndex->SetSounds(sound->second.clicks, "click-volume","selection-clicks");
-            }
-        }
-
-        Custom_OnLetGo = selection_release.M_Tab == 2;
-         if (selection_release.M_Tab == 0) {
-            auto list = ClickJson->GetMemeReleases();
-            auto sound = list.find(selection_release.Current_Sound_Meme);
-            if (sound != list.end()) {
-                ReleaseSoundIndex->SetSounds(sound->second.releases, "release-volume","selection-release");
-            }
-        } else {
-            auto list = ClickJson->GetUsefulReleases();
-            auto sound = list.find(selection_release.Current_Sound_Useful);
-            if (sound != list.end()) {
-                ReleaseSoundIndex->SetSounds(sound->second.releases, "release-volume","selection-release");
-            }
-        }
-
-    }
-}
-
 // the check to see if you should play the sound or not
 bool integrityCheck(PlayerObject* object, PlayerButton Pressed) {
     // play sounds when "only play on jump" settings is enabled and the player input is a jump, left movement, or right movement.
